# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "The Reward Loyalty API provides programmatic access to the Reward Loyalty platform, enabling integration with external systems, mobile apps, and custom workflows. This RESTful API supports authentication, loyalty card management, member operations, partner operations, and transaction processing.\n\n## Authentication\n\nAll authenticated endpoints use Bearer token authentication via Laravel Sanctum. Include your token in the Authorization header.\n\nTokens are obtained by calling the login endpoints for each user type (Admin, Partner, or Member).\n\n## Rate Limiting\n\nAPI requests are rate-limited to ensure fair usage. Standard limits apply per authenticated user.\n\n## Response Format\n\nAll responses are returned in JSON format. Successful responses include the requested data, while errors include a message field describing the issue.",
        "contact": {
            "name": "Reward Loyalty Website",
            "url": "https://distech.co.za"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/partner/cards/{cardUID}/{memberUID}/transactions/purchases": {
            "post": {
                "tags": [
                    "Partner"
                ],
                "summary": "Add a purchase to a card",
                "description": "Add a new purchase transaction to a card where the authenticated partner has access to.",
                "operationId": "addPurchase",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "cardUID",
                        "in": "path",
                        "description": "Unique identifier of the card to which the transaction will be added",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    },
                    {
                        "name": "memberUID",
                        "in": "path",
                        "description": "Unique identifier for the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Purchase data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "purchase_amount": {
                                        "description": "The amount of money spent on the purchase",
                                        "type": "number",
                                        "example": 100
                                    },
                                    "image": {
                                        "description": "The image associated with the transaction",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "description": "An optional note for the purchase",
                                        "type": "string",
                                        "example": "",
                                        "nullable": true
                                    },
                                    "staffId": {
                                        "description": "Staff ID",
                                        "type": "string",
                                        "format": "Kra8\\Snowflake\\HasSnowflakePrimary",
                                        "example": "50510833641460081"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/partner/cards/{cardUID}/{memberUID}/transactions/points": {
            "post": {
                "tags": [
                    "Partner"
                ],
                "summary": "Add points to a card",
                "description": "Add new points to a card where the authenticated partner has access to.",
                "operationId": "addPoints",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "cardUID",
                        "in": "path",
                        "description": "Unique identifier of the card to which the transaction will be added",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    },
                    {
                        "name": "memberUID",
                        "in": "path",
                        "description": "Unique identifier for the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Points data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "points": {
                                        "description": "The number of points to add",
                                        "type": "number",
                                        "example": 100
                                    },
                                    "image": {
                                        "description": "The image associated with the transaction",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "description": "An optional note for the points addition",
                                        "type": "string",
                                        "example": "",
                                        "nullable": true
                                    },
                                    "staffId": {
                                        "description": "Staff ID",
                                        "type": "string",
                                        "format": "Kra8\\Snowflake\\HasSnowflakePrimary",
                                        "example": "50510833641460081"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Points added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Response when authentication is required but not provided or invalid",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Authentication error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "description": "Response when the requested resource does not exist",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Not found error message",
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "description": "Response when request validation fails",
                "properties": {
                    "message": {
                        "description": "General validation error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Field-specific validation errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "description": "A completed loyalty card transaction",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "staff_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "member_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "card_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "partner_name": {
                        "type": "string"
                    },
                    "partner_email": {
                        "type": "string"
                    },
                    "purchase_amount": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "points_issued": {
                        "type": "integer"
                    },
                    "reward_redeemed": {
                        "type": "boolean"
                    },
                    "reward_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "transaction_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "admin_auth_token": {
                "type": "http",
                "description": "Administrator authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "partner_auth_token": {
                "type": "http",
                "description": "Partner authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Member authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}